docs: 完善demo和文档描述
