+++ docs/.vitepress/components/badge/base.vue
@@ -27,15 +27,6 @@
                 ]"
             />
         </FFormItem>
-        <FFormItem label="是否展示数据0:">
-            <FRadioGroup
-                v-model="showZero"
-                :options="[
-                    { label: '否(默认)', value: false },
-                    { label: '是', value: true },
-                ]"
-            />
-        </FFormItem>
     </FForm>
 
     <FDivider></FDivider>
@@ -81,21 +72,6 @@
             <FButton> Danger </FButton>
         </FBadge>
     </FSpace>
-
-    <FDivider></FDivider>
-
-    <FSpace>
-        <FBadge
-            :size="size"
-            :dot="dot"
-            :hidden="hidden"
-            :value="0"
-            :showZero="showZero"
-            class="item"
-        >
-            <FButton> Primary </FButton>
-        </FBadge>
-    </FSpace>
 </template>
 
 <script setup>
@@ -104,7 +80,6 @@ import { ref } from 'vue';
 const size = ref('middle');
 const dot = ref(false);
 const hidden = ref(false);
-const showZero = ref(false);
 </script>
 
 <style>
