+++ docs/.vitepress/components/badge/index.md
@@ -14,10 +14,17 @@ app.use(FBadge);
 
 ### 基础用法
 
-展示数值，小红点模式，显示隐藏，徽标类型，默认情况下数值 0 不展示
+展示数值，小红点模式，显示隐藏，徽标类型
 
 :::demo
 base.vue
+:::
+
+### 显示数值0
+默认情况下不展示数值0，如果 hidden为true，该属性不会生效，徽标仍会被隐藏
+
+:::demo
+showZero.vue
 :::
 
 ### 设定阈值
